Added simple WAV loader + sound effects

+++ assets/shaders/quad.vert
@@ -35,10 +35,10 @@ void main()
     transform.pos + transform.size                        // Bottom Right
     };
 
-    float left = transform.atlasOffset.x;
-    float top = transform.atlasOffset.y;
-    float right = transform.atlasOffset.x + transform.spriteSize.x;
-    float bottom = transform.atlasOffset.y + transform.spriteSize.y;
+    int left = transform.atlasOffset.x;
+    int top = transform.atlasOffset.y;
+    int right = transform.atlasOffset.x + transform.spriteSize.x;
+    int bottom = transform.atlasOffset.y + transform.spriteSize.y;
 
     vec2 textureCoords[6] =
     {
